LayoutTests/fullscreen/full-screen-restrictions-expected.txt

CONSOLE MESSAGE: Feature policy 'Fullscreen' check failed for iframe with origin '' and allow attribute 'fullscreen 'none''.
This tests the restrictions to entering full screen mode laid out in section 4.1 of the W3C Full Screen API
"The context object is not in a document."
EVENT(webkitfullscreenerror)
"The context object's node document, or an ancestor browsing context's document does not have the fullscreen enabled flag set."
EVENT(webkitfullscreenerror)
"A descendant browsing context's document has a non-empty fullscreen element stack."
EVENT(webkitfullscreenchange)
EVENT(webkitfullscreenerror)
EVENT(webkitfullscreenchange)
"This algorithm is not allowed to show a pop-up."
EVENT(webkitfullscreenerror)
END OF TEST
